fix(core): fix immediate generation; add conditional print to stdout

+++ src/main.cpp
@@ -10,10 +10,10 @@ int main(int argc, char* argv[]) {
     }
     std::string test_bin = argv[1];
     uint32_t base_address = std::stoul(argv[2], nullptr, 0);
-    uint32_t cycles = std::stoul(argv[3], nullptr, 0);
+    uint64_t cycles = std::stoul(argv[3], nullptr, 0);
     memory mem(base_address, test_bin);
     core rv32(base_address, &mem);
-    for (uint32_t i = 0; i < cycles; i++) rv32.exec();
+    for (uint64_t i = 0; i < cycles; i++) rv32.exec();
     rv32.dump();
     return 0;
 }
